feat: exported function and set loaders
